+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/CredentialFormatEnum.kt
@@ -1,6 +1,5 @@
 package at.asitplus.openid
 
-import at.asitplus.openid.CredentialFormatEnum.entries
 import kotlinx.serialization.Serializable
 
 @Serializable(with = CredentialFormatSerializer::class)
@@ -13,6 +12,12 @@ enum class CredentialFormatEnum(val text: String) {
     JWT_VC_JSON_LD("jwt_vc_json-ld"),
     JSON_LD("ldp_vc"),
     MSO_MDOC("mso_mdoc");
+
+    @Deprecated("To be removed along with VC_SD_JWT")
+    fun coerceDeprecations() = when(this) {
+        VC_SD_JWT -> DC_SD_JWT
+        else -> this
+    }
 
     companion object {
         fun parse(text: String) = entries.firstOrNull { it.text == text }